+++ mail/mail-common/src/user_context/initialization.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use crate::models::{ConversationCounters, MailSettings, MessageCounters, StoreLabelCounters};
+use crate::prefetch::Prefetch;
 use crate::{MailContextError, MailUserContext};
 use futures::future::try_join;
 use futures::try_join;
@@ -150,6 +151,13 @@ impl MailUserContext {
 
         debug!("Syncing Complete in {:?}", t0.elapsed());
         cb.on_stage(MailUserContextLoadingStage::Finished);
+
+        let (sender, receiver) = flume::unbounded();
+        let _ = ctx.prefetch.set(sender).inspect_err(|e| {
+            error!("Failed to set prefetch sender: {e:?}");
+        });
+        Prefetch::initialize(ctx.clone(), receiver).await;
+
         Ok(())
     }
 }